requirements.txt
coverage==4.5.1
Django==2.0
django-bootstrap-modal-forms==1.3.1
django-debug-toolbar==1.10.1
django-nose==1.4.5
django-widget-tweaks==1.4.3
nose==1.3.7
openfoodfacts==0.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.11.1
sqlparse==0.2.4
